+++ core-util/FunctionPointerBase.h
